✨ Feature: ArchiveHomeView, ArticleArchivingView 합치기
- 이든 광로 겹치는 테스크 합친 이후 데이터 연동
