+++ Projects/LastDance/LastDance/Sources/Common/Utils/Enum/Route.swift
@@ -27,5 +27,4 @@ enum Route: Hashable {
     case completeArticleList(selectedExhibitionId: String, selectedArtistId: Int)
     case artistReaction
     case artistReactionArchiveView
-    case archiveHome
 }
