+++ Projects/LastDance/LastDance/Sources/Domain/Model/Exhibition.swift
@@ -20,6 +20,7 @@ final class Exhibition {
     var coverImageName: String?
     var createdAt: String?
     var updatedAt: String?
+    var isUserSelected: Bool = false
 
     init(id: String,
          title: String,
@@ -29,7 +30,8 @@ final class Exhibition {
          venueId: Int? = nil,
          coverImageName: String? = nil,
          createdAt: String? = nil,
-         updatedAt: String? = nil) {
+         updatedAt: String? = nil,
+         isUserSelected: Bool = false) {
         self.id = id
         self.title = title
         self.descriptionText = descriptionText
@@ -39,5 +41,6 @@ final class Exhibition {
         self.coverImageName = coverImageName
         self.createdAt = createdAt
         self.updatedAt = updatedAt
+        self.isUserSelected = isUserSelected
     }
 }
