+++ Projects/LastDance/LastDance/Sources/Presentation/Archive/ViewModel/ArtArchiveViewModel.swift
@@ -1,0 +1,18 @@
+//
+//  ArtArchiveViewModel.swift
+//  LastDance
+//
+//  Created by 광로 on 10/20/25.
+//
+
+import SwiftUI
+
+struct ArtArchiveViewModel: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    ArtArchiveViewModel()
+}
