+++ Projects/LastDance/LastDance/Sources/Presentation/Archive/Views/ArtArchive.swift
@@ -1,0 +1,18 @@
+//
+//  ArtArchive.swift
+//  LastDance
+//
+//  Created by 광로 on 10/20/25.
+//
+
+import SwiftUI
+
+struct ArtArchive: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    ArtArchive()
+}
