+++ Projects/LastDance/LastDance/Sources/Presentation/ExhibitionDetail/Views/ExhibitionDetailView.swift
@@ -65,6 +65,7 @@ struct ExhibitionDetailView: View {
             if isArtist {
                 router.push(.artistReaction)
             } else {
+                viewModel.selectExhibitionAsUserExhibition()
                 router.push(.archive(id: exhibitionId))
             }
         }
